removed redundant casts

+++ src/org/pyload/android/client/fragments/AbstractPackageFragment.java
@@ -376,8 +376,7 @@ class PackageListAdapter extends BaseExpandableListAdapter {
 		if (pack.linkstotal == 0)
 			pack.linkstotal = 1;
 
-		holder.progress.setProgress((int) ((pack.linksdone * 100) / pack.links
-				.size()));
+		holder.progress.setProgress((pack.linksdone * 100) / pack.links	.size());
 		holder.size.setText(Utils.formatSize(pack.sizedone) + " / "
 				+ Utils.formatSize(pack.sizetotal));
 		holder.links.setText(pack.linksdone + " / " + pack.links.size());
